requirements.txt
requests==2.19.1
python-whois==0.7.0
